+++ components/camel-cassandraql/src/test/java/org/apache/camel/component/cassandra/integration/CassandraComponentProducerIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.cassandra;
+package org.apache.camel.component.cassandra.integration;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -28,6 +28,10 @@ import com.datastax.oss.driver.api.querybuilder.update.Update;
 import org.apache.camel.Produce;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.cassandra.CassandraConstants;
+import org.apache.camel.component.cassandra.CassandraEndpoint;
+import org.apache.camel.component.cassandra.MockLoadBalancingPolicy;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import static com.datastax.oss.driver.api.querybuilder.QueryBuilder.bindMarker;
@@ -36,7 +40,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class CassandraComponentProducerTest extends BaseCassandraTest {
+public class CassandraComponentProducerIT extends BaseCassandra {
 
     static final String CQL = "insert into camel_user(login, first_name, last_name) values (?, ?, ?)";
     static final String NO_PARAMETER_CQL = "select login, first_name, last_name from camel_user";
@@ -128,7 +132,7 @@ public class CassandraComponentProducerTest extends BaseCassandraTest {
         assertEquals("Claus 2", row.getString("first_name"));
         assertEquals("Ibsen 2", row.getString("last_name"));
 
-        assertTrue(MockLoadBalancingPolicy.used);
+        Assertions.assertTrue(MockLoadBalancingPolicy.used);
     }
 
     /**
