+++ components/camel-cassandraql/src/test/java/org/apache/camel/processor/aggregate/cassandra/CassandraAggregationIT.java
@@ -19,7 +19,7 @@ package org.apache.camel.processor.aggregate.cassandra;
 import org.apache.camel.AggregationStrategy;
 import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.cassandra.BaseCassandraTest;
+import org.apache.camel.component.cassandra.integration.BaseCassandra;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -27,7 +27,7 @@ import org.junit.jupiter.api.Test;
 /**
  * Unite test for {@link CassandraAggregationRepository}
  */
-public class CassandraAggregationTest extends BaseCassandraTest {
+public class CassandraAggregationIT extends BaseCassandra {
 
     private CassandraAggregationRepository aggregationRepository;
