+++ components/camel-cassandraql/src/test/java/org/apache/camel/processor/aggregate/cassandra/CassandraAggregationRepositoryIT.java
@@ -19,7 +19,7 @@ package org.apache.camel.processor.aggregate.cassandra;
 import java.util.Set;
 
 import org.apache.camel.Exchange;
-import org.apache.camel.component.cassandra.BaseCassandraTest;
+import org.apache.camel.component.cassandra.integration.BaseCassandra;
 import org.apache.camel.support.DefaultExchange;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -33,16 +33,14 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 /**
  * Unite test for {@link CassandraAggregationRepository}
  */
-public class NamedCassandraAggregationRepositoryTest extends BaseCassandraTest {
+public class CassandraAggregationRepositoryIT extends BaseCassandra {
 
     private CassandraAggregationRepository aggregationRepository;
 
     @Override
     protected void doPreSetup() throws Exception {
-        aggregationRepository = new NamedCassandraAggregationRepository(getSession(), "ID");
-        aggregationRepository.setTable("NAMED_CAMEL_AGGREGATION");
+        aggregationRepository = new CassandraAggregationRepository(getSession());
         aggregationRepository.start();
-
         super.doPreSetup();
     }
 
@@ -54,9 +52,7 @@ public class NamedCassandraAggregationRepositoryTest extends BaseCassandraTest {
     }
 
     private boolean exists(String key) {
-        return getSession().execute(String.format("select KEY from NAMED_CAMEL_AGGREGATION where NAME='ID' and KEY='%s'", key))
-                .one()
-               != null;
+        return getSession().execute(String.format("select KEY from CAMEL_AGGREGATION where KEY='%s'", key)).one() != null;
     }
 
     @Test
