+++ components/camel-cassandraql/src/test/java/org/apache/camel/processor/idempotent/cassandra/CassandraIdempotentIT.java
@@ -17,7 +17,7 @@
 package org.apache.camel.processor.idempotent.cassandra;
 
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.cassandra.BaseCassandraTest;
+import org.apache.camel.component.cassandra.integration.BaseCassandra;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -25,7 +25,7 @@ import org.junit.jupiter.api.Test;
 /**
  * Unite test for {@link CassandraIdempotentRepository}
  */
-public class CassandraIdempotentTest extends BaseCassandraTest {
+public class CassandraIdempotentIT extends BaseCassandra {
 
     private CassandraIdempotentRepository idempotentRepository;
